(PC-33428) feat(NC): move formatCurrencyFromCentsWithoutCurrencySymbol in shared/currency
