+++ src/features/search/pages/modals/PriceModal/PriceModal.tsx
@@ -16,8 +16,8 @@ import { MAX_PRICE_IN_CENTS } from 'features/search/helpers/reducer.helpers'
 import { priceSchema } from 'features/search/helpers/schema/priceSchema/priceSchema'
 import { SearchState } from 'features/search/types'
 import { useGetPacificFrancToEuroRate } from 'libs/firebase/firestore/exchangeRates/useGetPacificFrancToEuroRate'
-import { formatCurrencyFromCentsWithoutCurrencySymbol } from 'libs/parsers/formatCurrencyFromCentsWithoutCurrencySymbol'
 import { formatCurrencyFromCents } from 'shared/currency/formatCurrencyFromCents'
+import { formatCurrencyFromCentsWithoutCurrencySymbol } from 'shared/currency/formatCurrencyFromCentsWithoutCurrencySymbol'
 import { Currency, useGetCurrencyToDisplay } from 'shared/currency/useGetCurrencyToDisplay'
 import { useAvailableCredit } from 'shared/user/useAvailableCredit'
 import { InfoBanner } from 'ui/components/banners/InfoBanner'
