+++ src/shared/currency/formatCurrencyFromCentsWithoutCurrencySymbol.native.test.ts
