+++ src/shared/currency/formatCurrencyFromCentsWithoutCurrencySymbol.ts
